Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/origin-when-cross-origin/attr-referrer/cross-origin/http-http/a-tag/no-redirect/cross-origin.http-expected.txt



FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via a-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is cross-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/origin-when-cross-origin/attr-referrer/cross-origin/http-http/a-tag/no-redirect/cross-origin.http.html"
